third commit: code organization
